requirements.txt
requests==2.31.0
plotly==5.18.0
pandas==2.1.4
just-agents==0.4.0
python-dotenv==1.0.0
aiohttp==3.9.1
cachetools==5.3.2
dataclasses-json==0.6.3
